+++ CHANGELOG.md
@@ -10,6 +10,35 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ---
 
+## [2.0.6](https://github.com/FortAwesome/vue-fontawesome/releases/tag/2.0.6)  - 2021-10-18
+
+### Fixed
+* Include 1.3.0-beta versions in peer dependencies
+
+---
+
+## [2.0.5](https://github.com/FortAwesome/vue-fontawesome/releases/tag/2.0.5)  - 2021-10-15
+
+### Added
+* New sizes for `size` prop: 2xs, xl, 2xl
+* New animation booleans: `beat`, `fade`, `flash`, `spin-pulse`, & `spin-reverse`
+
+---
+
+## [2.0.4](https://github.com/FortAwesome/vue-fontawesome/releases/tag/2.0.4)  - 2021-10-15
+
+### Fixed
+* Relax the peer dependency for @fortawesome/fontawesome-svg-core
+
+---
+
+## [2.0.3](https://github.com/FortAwesome/vue-fontawesome/releases/tag/2.0.3)  - 2021-10-15
+
+### Fixed
+* Skip parse.icon if the icon is imported directly from an icon package
+
+---
+
 ## [2.0.2](https://github.com/FortAwesome/vue-fontawesome/releases/tag/2.0.2)  - 2020-12-17
 
 ### Fixed